updated DB schemas for PSQL to use 64-bit integers as indexes.
modified:
  data_model/postgres/8.3/06_drop_meta_alerts.sql
  data_model/postgres/8.3/11_create_wui.sql
  data_model/postgres/8.3/12_create_alerts.sql
  data_model/postgres/8.3/13_create_meta_alerts.sql
  data_model/postgres/8.3/14_create_processor_states.sql
  data_model/postgres/8.4/06_drop_meta_alerts.sql
  data_model/postgres/8.4/11_create_wui.sql
  data_model/postgres/8.4/12_create_alerts.sql
  data_model/postgres/8.4/13_create_meta_alerts.sql
  data_model/postgres/8.4/14_create_processor_states.sql

+++ data_model/postgres/8.3/11_create_wui.sql
@@ -6,7 +6,7 @@
 CREATE SEQUENCE wui_users_id_seq;
 CREATE TABLE    wui_users
 (
-  id int           PRIMARY KEY DEFAULT nextval('wui_users_id_seq'),
+  id      bigint   PRIMARY KEY DEFAULT nextval('wui_users_id_seq'),
   login   char(32) NOT NULL,
   pwdhash char(32) NOT NULL
 );
@@ -14,7 +14,7 @@ CREATE TABLE    wui_users
 -- logger_logs
 CREATE TABLE    logs
 (
-  id_wui_user int          NULL REFERENCES wui_users(id),
+  id_wui_user bigint       NULL REFERENCES wui_users(id),
   ctime       timestamp    NOT NULL DEFAULT now(),
   msg         varchar(128) NOT NULL
 );
